+++ Runtime/GameStartup/StartupConstants.cs
@@ -1,8 +1,15 @@
+using RKode.Utils;
+
 namespace RKode.Startup {
 public static class StartupConstants {
     public const string CONFIG_PATH = "RKode/ProjectConfig";
-    public const string CAN_LOAD_KEY = "RKode.Startup.CanLoadMain";
-    public const string PREVIOUS_SCENE_KEY = "RKode.Startup.PreviousScene";
-    public const string REDIRECT_COMPLETED_KEY = "RKode.Startup.RedirectCompleted";
+
+    public static readonly string CAN_LOAD_KEY = ProjectScopedKey.Get($"RKode.Startup.CanLoadMain");
+    public static readonly string PREVIOUS_SCENE_KEY = ProjectScopedKey.Get($"RKode.Startup.PreviousScene");
+    public static readonly string REDIRECT_COMPLETED_KEY = ProjectScopedKey.Get($"RKode.Startup.RedirectCompleted");
+
+	static StartupConstants() {
+		
+	}
 }
 }